mcp-servers/ASR/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
sse-starlette>=2.0.0
mcp>=0.1.0
fastmcp==2.9.0
requests==2.32.4
uuid==1.30
python-magic